Move all CDCT test files into their src directories This makes the location of the test files uniform and easier to find
